handle the case when current_uri is not available
On certain routes, like e.g. /index.php/installer, current_uri is not available in $context. If anyone hits such route, and the template includes ThirdParty plugin tags, the server responds with 500, not loading any template. If we try to get values from the $context using array_get, this error is prevented, since the value defaults to null if current_uri is not loaded.